subsume euclidean distance into human compactness abstract class
